Modified RA configuration files so that DPB limits not exceeded.
Modified the RA configuration files so that DPB limits not exceeded.
Also moved original RA configuration files to the cfg/misc folder, as
these can be used if coding a sequence where the picture size is half
that of the maximum for the level.

Corrected the field mode configuration files - they had tabs and
configuration lines without values.